requirements.txt
fastapi
uvicorn
pypdf
langchain
langchain-community
langchain-openai
python-dotenv
chromadb
tiktoken
streamlit
cryptography>=3.1 
